chore: rename file and some types
